requirements.txt
Django==1.9.4
diff-match-patch==20121119
django-crispy-forms==1.6.0
django-import-export==0.4.5
django-mptt==0.8.3
django-mptt-admin==0.3.6
gunicorn==19.4.5
pytz==2015.7
six==1.10.0
tablib==0.11.2
